Support more conversions, floating operations
Also upgraded the test harness to support generated multiple test
methods (so as to avoid overflowing any individual method) and to
handle translation of WAST floats that are not "normal".

  * implement comparison operators for floats (with the right G vs L!)
  * implement some conversions (this will eventually move to utils)
  * fixed some block layout bugs
  * fixed a bug in computing stack unwinds for `br`/`br_if`
  * `fNN.wast`, `fNN_bitwise.wast`, and `fNN_cmp.wast` all pass!

+++ lib/src/jvm/frame.rs
@@ -1067,7 +1067,7 @@ fn interpret_branch_instruction<Lbl, LblWide, LblNext>(
             }
         }
         LReturn => {
-            pop_offset_vec_expecting_type(stack, Integer)?;
+            pop_offset_vec_expecting_type(stack, Long)?;
             if *this_method_return_type != Some(FieldType::LONG) {
                 return Err(VerifierErrorKind::InvalidType);
             }
